+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+- Fails gracefully if property matchers (`property`, `enumerable` etc.) are used
+  on non-objects.
+
 ## 0.1.338 (Sep 26, 2013)
 - Fixes Must.js's own tests on Windows thanks to [Bart van der Schoor][bartvds].
 [bartvds]: https://github.com/Bartvds